improve phase delay convergence algorithm
If the link had previously converged to 0, use that as the
starting point for the guess.

The problem is that there is a race condition it seems between
convergence and the receiver deciding it's got a link problem --
if the convergence is too slow, the receiver will drop the link
and restart it again, forcing a full reconvergence.

If the link had previously converged, use the old value as
the starting point. This causes the link to converge much faster
the second time increasing the likelihood that the link will
"stick".

If the link structure was changed (cable swap, receiver swap)
then the initial convergence guess would of course be wrong, but
ideally the link converges again and resets the pre-set values.

The problem being not that the link doesn't converge -- the observed
behavior is that that it takes too long to converge, causing the
reset. e.g. it takes maybe 3/4 a second to converge, and 2 seconds
later the receiver resets the link to force a re-convergence. But
if the link can converge in under 1/4 second it doesn't force the
reset.