update MMCM methodology
couple bugs found:

1) power bits must be all 1's for DRP to work.
2) power register must be the first thing written
3) xapp888 seems to have a peculiar order of writing the registers

Instead of trying to create a table of human-readable
and modifiable constants, we just do a dump of the PLL
settings as created by vivado and jam the back in.

Compiled only for 1080p mode so far, and low/high bandwidth
settings.

Other modes use the legacy method.