refactor tester output to be JSON formatted
This is all going into a test program anyways and ultimately
rendered via a browser to a UI if anything.

JSON output has been validated using
https://jsonformatter.curiousconcept.com/

And there is also a "force error" #define added to try and
catch as many of the error messages as possible to confirm
their JSON compatibility.