add support for hardware link training algorithm
There's now an CSR that can specify which algorithm to use:
0 = original software algorithm
1 = cheesy hack to confirm ==/!= parity problem
2 = hardware-based algorithm

When picking the hardware-based algorithm you also need
to specify how many bit times wide (in terms of delay units)
the eye is.

This code should probably be refactor to derive the bit time
from the known delay clock reference that's given, but we're
in a hurry to meet a MVP image date right now.