Handle case where km == 0 with a hot plug retry
For some reason on very first boot, the first key exchange
always comes back with km == 0. This doesn't happen later on
even when changing receivers (so it's not an issue of the firmware
being unable to catch the hot plug exchange and factor the keys
in time).

I suspect it has to do with all the churn going on with link
training and interrupts firing like crazy on boot which causes
an event to get missed somewhere. Anyways, handling the case
where the Km is observed to be bad is a simple fix to the
problem.