Add alt_delay channel bonding mechanism
Original channel bonding has the property of changing the relative
skew of channels every time a symbol error is encountered in
the control stream. This can lead to downstream monitors having
a hard time locking on to a signal.

The alternate channel bonding mechanism here finds an alignment
and sticks with it (in part because the phase aligner also does
the same).

This noticeably improves image stability on links with high
bit error rates